requirements.txt

flask==3.1.1
flask-cors==4.0.0
openai==1.91.0
python-dotenv==1.1.1
gunicorn==23.0.0
markdown==3.6
